Downloads/kajal/orchestrator/requirements.txt
pipecat-ai[anthropic,silero,twilio]==0.0.47
anthropic==0.25.0
twilio==9.0.4
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.4.2
numpy==1.26.4
aiohttp==3.9.5
loguru==0.7.2
